split main up into different files to hopefully make things more scaleable.
